+++ pytao/constraints/config.py
@@ -4,67 +4,8 @@ from typing import Annotated, Literal, Union
 from pydantic import BaseModel, Field
 
 from pytao.constraints.observables import IsCloseResult, Observable, Observation
-from pytao.constraints.observables.datum import DatumIsClose, DatumIsCloseResult, DatumObservable, DatumObservation
-from pytao.model import ElementFloor, ElementFloorAll, ElementFloorPosition, ElementOrbit, ElementTwiss
-from pytao.model.ele.ele import Element
-
-from pytao.constraints.observables.ele import EleIsClose, EleIsCloseResult, EleObservable, EleObservation
-
-
-class EleLiteral(BaseModel):
-    beta_a: float | None = None
-    alpha_a: float | None = None
-    beta_b: float | None = None
-    alpha_b: float | None = None
-    eta_x: float | None = None
-    etap_x: float | None = None
-    eta_y: float | None = None
-    etap_y: float | None = None
-    p0c: float | None = None
-    floor_x: float | None = None
-    floor_y: float | None = None
-    floor_z: float | None = None
-
-    def to_observation(self) -> EleObservation:
-        new_twiss = ElementTwiss(
-            **{k: v for k, v in [
-                ("beta_a", self.beta_a), ("alpha_a", self.alpha_a),
-                ("beta_b", self.beta_b), ("alpha_b", self.alpha_b),
-                ("eta_x", self.eta_x), ("etap_x", self.etap_x),
-                ("eta_y", self.eta_y), ("etap_y", self.etap_y),
-            ] if v is not None}
-        )
-        new_orbit = ElementOrbit(
-            **{k: v for k, v in [("p0c", self.p0c)] if v is not None}
-        )
-        dummy_floor = ElementFloor(which="model", where="end", actual=None, reference=None, slaves={})
-        new_floor = ElementFloorAll(
-            which="model",
-            beginning=dummy_floor,
-            center=dummy_floor,
-            end=ElementFloor(
-                which="model",
-                where="end",
-                actual=ElementFloorPosition.model_construct(
-                    **{k: v for k, v in [
-                        ("x", self.floor_x), ("y", self.floor_y), ("z", self.floor_z)
-                    ] if v is not None}
-                ),
-                reference=None,
-                slaves={},
-            ),
-        )
-        
-        # Note: use model_construct to pack values into model without validation so we may partially populate for
-        # use in the constraint checking code.
-        element = Element.model_construct(
-            ele_id="_literal",
-            which="model",
-            twiss=new_twiss,
-            orbit=new_orbit,
-            floor=new_floor,
-        )
-        return EleObservation(element=element)
+from pytao.constraints.observables.datum import DatumIsClose, DatumIsCloseResult, DatumObservable, DatumObservation, DatumLiteral
+from pytao.constraints.observables.ele import EleIsClose, EleIsCloseResult, EleObservable, EleObservation, EleLiteral
 
 
 class EqualityConstraint(BaseModel):
@@ -136,17 +77,6 @@ class EleLiteralEquality(EqualityConstraint):
         if not isinstance(obs_a, EleObservation):
             raise TypeError(f"expected EleObservation, got {type(obs_a)}")
         return self.comparison(obs_a, self.expected.to_observation())
-
-
-class DatumLiteral(BaseModel):
-    model_value: float | None = None
-    design_value: float | None = None
-
-    def to_observation(self, base: DatumObservation) -> DatumObservation:
-        return DatumObservation(
-            model_value=self.model_value if self.model_value is not None else base.model_value,
-            design_value=self.design_value if self.design_value is not None else base.design_value,
-        )
 
 
 class DatumLiteralEquality(EqualityConstraint):
