+++ pytao/constraints/observables/datum.py
@@ -1,6 +1,6 @@
 from typing import Literal
 
-from pydantic import Field
+from pydantic import Field, BaseModel
 
 from pytao import Tao
 from pytao.constraints.observables.base import CheckResult, IsClose, IsCloseResult, Observable, Observation
@@ -47,6 +47,17 @@ class DatumObservable(Observable):
         return DatumObservation(
             model_value=result["model_value"],
             design_value=result["design_value"],
+        )
+
+
+class DatumLiteral(BaseModel):
+    model_value: float | None = None
+    design_value: float | None = None
+
+    def to_observation(self, base: DatumObservation) -> DatumObservation:
+        return DatumObservation(
+            model_value=self.model_value if self.model_value is not None else base.model_value,
+            design_value=self.design_value if self.design_value is not None else base.design_value,
         )
