+++ pytao/constraints/observables/ele.py
@@ -6,6 +6,8 @@ from pytao import Tao
 from pytao.model import Element
 from pytao.constraints.observables.base import CheckResult, IsClose, IsCloseResult, Observable, Observation
 from pytao.constraints.observables.twiss import BmagTwissComparison, twiss_comparison_types
+from pytao.model import ElementFloor, ElementFloorAll, ElementFloorPosition, ElementOrbit, ElementTwiss
+from pytao.model.ele.ele import Element
 
 
 class EleObservation(Observation):
@@ -24,6 +26,62 @@ class EleObservable(Observable):
 
     def __call__(self, tao: Tao) -> EleObservation:
         return EleObservation(element=tao.ele(self.ele_id))
+
+
+class EleLiteral(BaseModel):
+    beta_a: float | None = None
+    alpha_a: float | None = None
+    beta_b: float | None = None
+    alpha_b: float | None = None
+    eta_x: float | None = None
+    etap_x: float | None = None
+    eta_y: float | None = None
+    etap_y: float | None = None
+    p0c: float | None = None
+    floor_x: float | None = None
+    floor_y: float | None = None
+    floor_z: float | None = None
+
+    def to_observation(self) -> EleObservation:
+        new_twiss = ElementTwiss(
+            **{k: v for k, v in [
+                ("beta_a", self.beta_a), ("alpha_a", self.alpha_a),
+                ("beta_b", self.beta_b), ("alpha_b", self.alpha_b),
+                ("eta_x", self.eta_x), ("etap_x", self.etap_x),
+                ("eta_y", self.eta_y), ("etap_y", self.etap_y),
+            ] if v is not None}
+        )
+        new_orbit = ElementOrbit(
+            **{k: v for k, v in [("p0c", self.p0c)] if v is not None}
+        )
+        dummy_floor = ElementFloor(which="model", where="end", actual=None, reference=None, slaves={})
+        new_floor = ElementFloorAll(
+            which="model",
+            beginning=dummy_floor,
+            center=dummy_floor,
+            end=ElementFloor(
+                which="model",
+                where="end",
+                actual=ElementFloorPosition.model_construct(
+                    **{k: v for k, v in [
+                        ("x", self.floor_x), ("y", self.floor_y), ("z", self.floor_z)
+                    ] if v is not None}
+                ),
+                reference=None,
+                slaves={},
+            ),
+        )
+        
+        # Note: use model_construct to pack values into model without validation so we may partially populate for
+        # use in the constraint checking code.
+        element = Element.model_construct(
+            ele_id="_literal",
+            which="model",
+            twiss=new_twiss,
+            orbit=new_orbit,
+            floor=new_floor,
+        )
+        return EleObservation(element=element)
 
 
 class TolComparison(BaseModel):
